pyproject.toml
[tool.poetry]
name = "nautobot-ssot"
version = "1.2.0"
description = "Nautobot Single Source of Truth"
authors = ["Network to Code, LLC <opensource@networktocode.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/nautobot/nautobot-plugin-ssot"
repository = "https://github.com/nautobot/nautobot-plugin-ssot"
documentation = "https://nautobot-plugin-ssot.readthedocs.io"
keywords = ["nautobot", "nautobot-plugin"]
classifiers = [
    "Intended Audience :: Developers",
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
include = [
    "LICENSE",
    "README.md",
]
packages = [
    { include = "nautobot_ssot" },
]

[tool.poetry.dependencies]
python = "^3.7"
nautobot = "^1.4"
diffsync = "^1.8.0"
Markdown = "!=3.3.5"

[tool.poetry.dev-dependencies]
invoke = "*"
black = "*"
django-debug-toolbar = "*"
yamllint = "*"
bandit = "*"
pylint = "*"
pylint-django = "*"
pydocstyle = "*"
flake8 = "~3.9.1"
coverage = "*"
mkdocs = "*"


[tool.poetry.plugins."nautobot_ssot.data_sources"]
"example" = "nautobot_ssot.sync.example:ExampleSyncWorker"

[tool.poetry.plugins."nautobot_ssot.data_targets"]
"example" = "nautobot_ssot.sync.example:ExampleSyncWorker"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
# "too-few-public-methods" is just plain noise.
disable = """,
    line-too-long,
    bad-continuation,
    too-few-public-methods,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "-vv --doctest-modules"
